requirements.txt
pyyaml>=6.0.1
matplotlib>=3.5.0
numpy>=1.20.0
graphviz>=0.19.0 